requirements.txt
Django==1.9.7
djangorestframework==3.4.0
Markdown==2.6.6
django-filter==0.13.0
djoser==0.5.0
django-custom-user==0.6
